[TASK] Apply default content sliding for not existing colPoses
For colPoses that does not exist in backend layout automatically apply sliding from parent (with "-1" value).

Useful in cases when e.g. footer is inherited by it is not editable on the child page (in previous solution it was not slided until same colPos was added into backend layout of the child page).